Add custom Lua searcher to fix locals detection.
In some cases the line hook will not pick up the last defined local
variable if that definition is the last statement of a chunk. For
modules using the deprecated `module` function that situation could
occur, so all Lua module chunks are loaded with an explicit `return`
statement in the replacement Lua searcher.